+++ CHANGES.md
@@ -1,4 +1,6 @@
 1. Change mod key to super
 change mod1mask to mod4mask in config.h
 
-2.
+2. Change font to 18
+
+3. Change kill window to ctrl shift q